move the fetch module to a separate file
