+++ apollo-router/src/query_planner/fetch.rs
@@ -1,0 +1,360 @@
+use std::collections::HashMap;
+use std::fmt::Display;
+use std::sync::Arc;
+
+use indexmap::IndexSet;
+use serde::Deserialize;
+use serde::Serialize;
+use tower::ServiceExt;
+use tracing::instrument;
+use tracing::Instrument;
+
+use super::selection::select_object;
+use super::selection::Selection;
+use super::ExecutionParameters;
+use crate::error::Error;
+use crate::error::FetchError;
+use crate::graphql::Request;
+use crate::json_ext::Object;
+use crate::json_ext::Path;
+use crate::json_ext::Value;
+use crate::json_ext::ValueExt;
+use crate::services::subgraph_service::SubgraphServiceFactory;
+use crate::*;
+
+/// GraphQL operation type.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+#[non_exhaustive]
+pub enum OperationKind {
+    Query,
+    Mutation,
+    Subscription,
+}
+
+impl Display for OperationKind {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            OperationKind::Query => write!(f, "Query"),
+            OperationKind::Mutation => write!(f, "Mutation"),
+            OperationKind::Subscription => write!(f, "Subscription"),
+        }
+    }
+}
+
+impl Default for OperationKind {
+    fn default() -> Self {
+        OperationKind::Query
+    }
+}
+
+/// A fetch node.
+#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct FetchNode {
+    /// The name of the service or subgraph that the fetch is querying.
+    pub(crate) service_name: String,
+
+    /// The data that is required for the subgraph fetch.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    #[serde(default)]
+    pub(crate) requires: Vec<Selection>,
+
+    /// The variables that are used for the subgraph fetch.
+    pub(crate) variable_usages: Vec<String>,
+
+    /// The GraphQL subquery that is used for the fetch.
+    pub(crate) operation: String,
+
+    /// The GraphQL subquery operation name.
+    pub(crate) operation_name: Option<String>,
+
+    /// The GraphQL operation kind that is used for the fetch.
+    pub(crate) operation_kind: OperationKind,
+
+    /// Optional id used by Deferred nodes
+    pub(crate) id: Option<String>,
+}
+
+struct Variables {
+    variables: Object,
+    paths: HashMap<Path, usize>,
+}
+
+impl Variables {
+    #[instrument(skip_all, level = "debug", name = "make_variables")]
+    async fn new(
+        requires: &[Selection],
+        variable_usages: &[String],
+        data: &Value,
+        current_dir: &Path,
+        request: &Arc<http::Request<Request>>,
+        schema: &Schema,
+        enable_deduplicate_variables: bool,
+    ) -> Option<Variables> {
+        let body = request.body();
+        if !requires.is_empty() {
+            let mut variables = Object::with_capacity(1 + variable_usages.len());
+
+            variables.extend(variable_usages.iter().filter_map(|key| {
+                body.variables
+                    .get_key_value(key.as_str())
+                    .map(|(variable_key, value)| (variable_key.clone(), value.clone()))
+            }));
+
+            let mut paths: HashMap<Path, usize> = HashMap::new();
+            let (paths, representations) = if enable_deduplicate_variables {
+                let mut values: IndexSet<Value> = IndexSet::new();
+                data.select_values_and_paths(current_dir, |path, value| {
+                    if let Value::Object(content) = value {
+                        if let Ok(Some(value)) = select_object(content, requires, schema) {
+                            match values.get_index_of(&value) {
+                                Some(index) => {
+                                    paths.insert(path.clone(), index);
+                                }
+                                None => {
+                                    paths.insert(path.clone(), values.len());
+                                    values.insert(value);
+                                }
+                            }
+                        }
+                    }
+                });
+
+                if values.is_empty() {
+                    return None;
+                }
+
+                (paths, Value::Array(Vec::from_iter(values)))
+            } else {
+                let mut values: Vec<Value> = Vec::new();
+                data.select_values_and_paths(current_dir, |path, value| {
+                    if let Value::Object(content) = value {
+                        if let Ok(Some(value)) = select_object(content, requires, schema) {
+                            paths.insert(path.clone(), values.len());
+                            values.push(value);
+                        }
+                    }
+                });
+
+                if values.is_empty() {
+                    return None;
+                }
+
+                (paths, Value::Array(Vec::from_iter(values)))
+            };
+            variables.insert("representations", representations);
+
+            Some(Variables { variables, paths })
+        } else {
+            // with nested operations (Query or Mutation has an operation returning a Query or Mutation),
+            // when the first fetch fails, the query plan will still execute up until the second fetch,
+            // where `requires` is empty (not a federated fetch), the current dir is not emmpty (child of
+            // the previous operation field) and the data is null. In that case, we recognize that we
+            // should not perform the next fetch
+            if !current_dir.is_empty()
+                && data
+                    .get_path(current_dir)
+                    .map(|value| value.is_null())
+                    .unwrap_or(true)
+            {
+                return None;
+            }
+
+            Some(Variables {
+                variables: variable_usages
+                    .iter()
+                    .filter_map(|key| {
+                        body.variables
+                            .get_key_value(key.as_str())
+                            .map(|(variable_key, value)| (variable_key.clone(), value.clone()))
+                    })
+                    .collect::<Object>(),
+                paths: HashMap::new(),
+            })
+        }
+    }
+}
+
+impl FetchNode {
+    #[allow(clippy::too_many_arguments)]
+    pub(crate) async fn fetch_node<'a, SF>(
+        &'a self,
+        parameters: &'a ExecutionParameters<'a, SF>,
+        data: &'a Value,
+        current_dir: &'a Path,
+    ) -> Result<(Value, Vec<Error>), FetchError>
+    where
+        SF: SubgraphServiceFactory,
+    {
+        let FetchNode {
+            operation,
+            operation_kind,
+            operation_name,
+            service_name,
+            ..
+        } = self;
+
+        let Variables { variables, paths } = match Variables::new(
+            &self.requires,
+            self.variable_usages.as_ref(),
+            data,
+            current_dir,
+            // Needs the original request here
+            parameters.supergraph_request,
+            parameters.schema,
+            parameters.options.enable_deduplicate_variables,
+        )
+        .await
+        {
+            Some(variables) => variables,
+            None => {
+                return Ok((Value::Object(Object::default()), Vec::new()));
+            }
+        };
+
+        let subgraph_request = SubgraphRequest::builder()
+            .supergraph_request(parameters.supergraph_request.clone())
+            .subgraph_request(
+                http_ext::Request::builder()
+                    .method(http::Method::POST)
+                    .uri(
+                        parameters
+                            .schema
+                            .subgraphs()
+                            .find_map(|(name, url)| (name == service_name).then(|| url))
+                            .unwrap_or_else(|| {
+                                panic!(
+                                    "schema uri for subgraph '{}' should already have been checked",
+                                    service_name
+                                )
+                            })
+                            .clone(),
+                    )
+                    .body(
+                        Request::builder()
+                            .query(operation)
+                            .and_operation_name(operation_name.clone())
+                            .variables(variables.clone())
+                            .build(),
+                    )
+                    .build()
+                    .expect("it won't fail because the url is correct and already checked; qed"),
+            )
+            .operation_kind(*operation_kind)
+            .context(parameters.context.clone())
+            .build();
+
+        let service = parameters
+            .service_factory
+            .new_service(service_name)
+            .expect("we already checked that the service exists during planning; qed");
+
+        // TODO not sure if we need a RouterReponse here as we don't do anything with it
+        let (_parts, response) = service
+            .oneshot(subgraph_request)
+            .instrument(tracing::trace_span!("subfetch_stream"))
+            .await
+            // TODO this is a problem since it restores details about failed service
+            // when errors have been redacted in the include_subgraph_errors module.
+            // Unfortunately, not easy to fix here, because at this point we don't
+            // know if we should be redacting errors for this subgraph...
+            .map_err(|e| FetchError::SubrequestHttpError {
+                service: service_name.to_string(),
+                reason: e.to_string(),
+            })?
+            .response
+            .into_parts();
+
+        super::log::trace_subfetch(service_name, operation, &variables, &response);
+
+        if !response.is_primary() {
+            return Err(FetchError::SubrequestUnexpectedPatchResponse {
+                service: service_name.to_owned(),
+            });
+        }
+
+        // fix error path and erase subgraph error messages (we cannot expose subgraph information
+        // to the client)
+        let errors: Vec<Error> = response
+            .errors
+            .into_iter()
+            .map(|error| Error {
+                locations: error.locations,
+                path: error.path.map(|path| current_dir.join(path)),
+                message: error.message,
+                extensions: error.extensions,
+            })
+            .collect();
+
+        match self.response_at_path(current_dir, paths, response.data.unwrap_or_default()) {
+            Ok(value) => {
+                if let Some(id) = &self.id {
+                    if let Some(sender) = parameters.deferred_fetches.get(id.as_str()) {
+                        if let Err(e) = sender.clone().send((value.clone(), errors.clone())) {
+                            tracing::error!("error sending fetch result at path {} and id {:?} for deferred response building: {}", current_dir, self.id, e);
+                        }
+                    }
+                }
+
+                Ok((value, errors))
+            }
+            Err(e) => Err(e),
+        }
+    }
+
+    #[instrument(skip_all, level = "debug", name = "response_insert")]
+    fn response_at_path<'a>(
+        &'a self,
+        current_dir: &'a Path,
+        paths: HashMap<Path, usize>,
+        data: Value,
+    ) -> Result<Value, FetchError> {
+        if !self.requires.is_empty() {
+            // we have to nest conditions and do early returns here
+            // because we need to take ownership of the inner value
+            if let Value::Object(mut map) = data {
+                if let Some(entities) = map.remove("_entities") {
+                    tracing::trace!("received entities: {:?}", &entities);
+
+                    if let Value::Array(array) = entities {
+                        let mut value = Value::default();
+
+                        for (path, entity_idx) in paths {
+                            value.insert(
+                                &path,
+                                array
+                                    .get(entity_idx)
+                                    .ok_or_else(|| FetchError::ExecutionInvalidContent {
+                                        reason: "Received invalid content for key `_entities`!"
+                                            .to_string(),
+                                    })?
+                                    .clone(),
+                            )?;
+                        }
+                        return Ok(value);
+                    } else {
+                        return Err(FetchError::ExecutionInvalidContent {
+                            reason: "Received invalid type for key `_entities`!".to_string(),
+                        });
+                    }
+                }
+            }
+
+            Err(FetchError::ExecutionInvalidContent {
+                reason: "Missing key `_entities`!".to_string(),
+            })
+        } else {
+            Ok(Value::from_path(current_dir, data))
+        }
+    }
+
+    #[cfg(test)]
+    pub(crate) fn service_name(&self) -> &str {
+        &self.service_name
+    }
+
+    pub(crate) fn operation_kind(&self) -> &OperationKind {
+        &self.operation_kind
+    }
+}
